refactor: rename components/classes
