+++ src/temp/MapPreview/MapPreview.module.scss
@@ -1,5 +1,5 @@
 @import '../../theme/theme';
-.diveIntoMapMainContainer {
+.mapPreviewContainer {
   max-width: 306px;
   width: 100%;
   height: 341px;
@@ -10,7 +10,7 @@
   justify-content: center;
 }
 
-.diveIntoProjectButton {
+.showMapButton {
   display: flex;
   justify-content: center;
   align-items: center;
@@ -29,6 +29,6 @@
   }
 }
 
-.diveIconContainer {
+.buttonIconContainer {
   margin-top: 1px;
 }
