+++ src/temp/MapPreview/ShowMapButton.tsx
@@ -1,14 +1,14 @@
-import DiveIcon from '../icons/DiveIcon';
-import style from './DiveIntoMap.module.scss';
+import ExpandIcon from '../icons/ExpandIcon';
+import style from './MapPreview.module.scss';
 import { useTranslation } from 'next-i18next';
 
 const ShowMapButton = () => {
   const { t } = useTranslation(['projectDetails']);
   return (
     <>
-      <button className={style.diveIntoProjectButton}>
-        <div className={style.diveIconContainer}>
-          <DiveIcon />
+      <button className={style.showMapButton}>
+        <div className={style.buttonIconContainer}>
+          <ExpandIcon />
         </div>
         <div className={style.label}>
           {t('projectDetails:diveIntoTheProject')}
