+++ src/temp/icons/ExpandIcon.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-const DiveIcon = () => (
+const ExpandIcon = () => (
   <svg xmlns="http://www.w3.org/2000/svg" width={16} height={17} fill="none">
     <path
       fill="#fff"
@@ -7,4 +7,4 @@ const DiveIcon = () => (
     />
   </svg>
 );
-export default DiveIcon;
+export default ExpandIcon;
